Refactor internal src dir
