+++ include/dlgcpp/align.h
@@ -1,0 +1,18 @@
+#pragma once
+
+namespace dlgcpp
+{
+    enum class HorizontalAlign
+    {
+        Left = 0,
+        Center,
+        Right
+    };
+
+    enum class VerticalAlign
+    {
+        Top = 0,
+        Center,
+        Bottom
+    };
+}
