+++ src/win32/CMakeLists.txt
@@ -1,0 +1,91 @@
+cmake_minimum_required(VERSION 3.27)
+
+set(platform_src
+    control.cpp
+    control_p.h
+    dialog.cpp
+    dialog_p.h
+    event_p.h
+    menu.cpp
+    menu_p.h
+    controls/button.cpp
+    controls/button_p.h
+    controls/checkbox.cpp
+    controls/checkbox_p.h
+    controls/combobox.cpp
+    controls/combobox_p.h
+    controls/custom.cpp
+    controls/custom_p.h
+    controls/image.cpp
+    controls/image_p.h
+    controls/label.cpp
+    controls/label_p.h
+    controls/listbox.cpp
+    controls/listbox_p.h
+    controls/listview.cpp
+    controls/listview_col.cpp
+    controls/listview_p.h
+    controls/option_btn.cpp
+    controls/option_btn_p.h
+    controls/progress.cpp
+    controls/progress_p.h
+    controls/slider.cpp
+    controls/slider_imp.cpp
+    controls/slider_imp.h
+    controls/slider_p.h
+    controls/tab_item.cpp
+    controls/tabs.cpp
+    controls/tabs_p.h
+    controls/textbox.cpp
+    controls/textbox_p.h
+    controls/toolbar.cpp
+    controls/toolbar_item.cpp
+    controls/toolbar_p.h
+    controls/trackbar.cpp
+    controls/trackbar_p.h
+    controls/treeview.cpp
+    controls/treeview_node.cpp
+    controls/treeview_p.h
+    dialogs/about.cpp
+    dialogs/about_p.h
+    dialogs/file.cpp
+    dialogs/file_p.h
+    dialogs/folder.cpp
+    dialogs/folder_p.h
+    dialogs/message.cpp
+    dialogs/message_p.h
+    dialogs/property.cpp
+    dialogs/property_p.h
+    dialogs/splash.cpp
+    dialogs/splash_p.h
+    gfx/context.cpp
+    gfx/context.h
+    utility/font.cpp
+    utility/font.h
+    utility/image.cpp
+    utility/image.h
+    utility/keys.cpp
+    utility/keys.h
+    utility/message.cpp
+    utility/message.h
+    utility/string.cpp
+    utility/string.h
+    utility/units.cpp
+    utility/units.h
+)
+
+# prefix paths
+list(TRANSFORM platform_src PREPEND "${CMAKE_CURRENT_SOURCE_DIR}/")
+
+set(platform_includes
+    ${CMAKE_CURRENT_SOURCE_DIR}
+)
+
+set(platform_deps
+    comctl32
+    gdiplus
+)
+
+set(platform_src ${platform_src} PARENT_SCOPE)
+set(platform_includes ${platform_includes} PARENT_SCOPE)
+set(platform_deps ${platform_deps} PARENT_SCOPE)
